更新模型设计,并解决因此导致问题

+++ only4-KSP-cap4j-application/src/main/java/com/only4/application/commands/article/UpdateArticleCommentCountCmd.java
@@ -50,7 +50,7 @@ public class UpdateArticleCommentCountCmd {
     @AllArgsConstructor
     public static class Request implements RequestParam<Response> {
         Long articleId;
-        Long CommentCount;
+        Integer CommentCount;
     }
 
     /**
